PT-1987-projects: Changed project to be a group of patients.
